requirements.txt
Flask==1.1.2
gunicorn==20.0.4
requests==2.24.0
google-cloud-firestore==1.8.1
mailjet-rest==1.3.3
google-cloud-secret-manager==1.0.0
